backend/services/deep-research/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.99.5
python-dotenv==1.1.1
requests==2.32.5
